fix : Moved matchmaker Allocation into the Multiplay SQP script.
Instead of taking a dependency on Multiplay in the server manager and calling the Multiplay SDK there, I decided to move the functionality into the SQP Service class.
We can discuss if there is a better name for it now that it does more than just SQP? I'm fine with it for now.